+++ run/get-git-diff-files.go
@@ -1,0 +1,21 @@
+package run
+
+import (
+	"eslint-git-diff/structs"
+	"fmt"
+	"os/exec"
+	"strings"
+)
+
+func GetGitDiffFiles(flg structs.Flg) string {
+	gitArguments := strings.Fields("diff --name-only --diff-filter=d " + flg.Branch)
+
+	out, err := exec.Command("git", gitArguments...).CombinedOutput()
+
+	if err != nil {
+		fmt.Printf("GIT DIFF ERROR!\n%s\n", string(out))
+		panic(err)
+	}
+
+	return string(out)
+}
