+++ structs/flg.go
